+++ src/GitHub_Updater/Theme.php
@@ -254,11 +254,9 @@ class Theme extends Base {
 			}
 
 			/*
-			 * Remove WordPress update row in theme row, only in multisite.
 			 * Add update row to theme row, only in multisite.
 			 */
 			if ( is_multisite() ) {
-				add_action( 'after_theme_row', array( &$this, 'remove_after_theme_row' ), 10, 2 );
 				if ( ! $this->tag ) {
 					add_action( "after_theme_row_$theme->repo", array( &$this, 'wp_theme_update_row' ), 10, 2 );
 					add_action( "after_theme_row_$theme->repo", array( &$this, 'theme_branch_switcher' ), 10, 2 );
@@ -273,6 +271,8 @@ class Theme extends Base {
 	 * Load pre-update filters.
 	 */
 	public function load_pre_filters() {
+		wp_enqueue_style( 'github-updater', plugins_url( basename( dirname( dirname( __DIR__ ) ) ) ) . '/css/github-updater.css' );
+
 		if ( ! is_multisite() ) {
 			add_filter( 'wp_prepare_themes_for_js', array( &$this, 'customize_theme_update_html' ) );
 		}
@@ -318,39 +318,12 @@ class Theme extends Base {
 				$response->last_updated = $theme->last_updated;
 				$response->rating       = $theme->rating;
 				$response->num_ratings  = $theme->num_ratings;
-				if ( $theme->private ) {
-					add_action( 'admin_head', array( $this, 'remove_rating_in_private_repo' ) );
-				}
+
 				break;
 			}
 		}
-		add_action( 'admin_head', array( $this, 'fix_display_in_themes_api' ) );
 
 		return $response;
-	}
-
-	/**
-	 * Fix for new issue in 3.9 :-(
-	 */
-	public function fix_display_in_themes_api() {
-		?>
-		<style>
-			#theme-installer div.install-theme-info {
-				display: block !important;
-			}
-
-			#theme-installer.wp-full-overlay.single-theme, .wp-full-overlay-sidebar {
-				position: relative;
-			}
-		</style>
-		<?php
-	}
-
-	/**
-	 * Remove star rating for private themes.
-	 */
-	public function remove_rating_in_private_repo() {
-		echo '<style> #theme-installer div.install-theme-info div.star-rating { display: none; } </style>';
 	}
 
 	/**
@@ -362,28 +335,24 @@ class Theme extends Base {
 	 * @author Seth Carstens
 	 */
 	public function wp_theme_update_row( $theme_key, $theme ) {
+		$current = get_site_transient( 'update_themes' );
 
-		$current            = get_site_transient( 'update_themes' );
-		$themes_allowedtags = array(
-			'a'       => array( 'href' => array(), 'title' => array() ),
-			'abbr'    => array( 'title' => array() ),
-			'acronym' => array( 'title' => array() ),
-			'code'    => array(),
-			'em'      => array(),
-			'strong'  => array(),
-		);
-		$theme_name         = wp_kses( $theme['Name'], $themes_allowedtags );
-		$wp_list_table      = _get_list_table( 'WP_MS_Themes_List_Table' );
-		$install_url        = self_admin_url( "theme-install.php" );
-		$details_url        = esc_attr( add_query_arg(
-			array(
-				'tab'       => 'theme-information',
-				'theme'     => $theme_key,
-				'TB_iframe' => 'true',
-				'width'     => 270,
-				'height'    => 400,
-			),
-			$install_url ) );
+		add_filter( 'su_modify_theme_details_url', function ( $details_url, $theme_key ) {
+			$themes = Theme::instance()->config;
+			if ( array_key_exists( $theme_key, $themes ) ) {
+				$details_url = esc_attr( add_query_arg(
+					array(
+						'tab'       => 'theme-information',
+						'theme'     => $theme_key,
+						'TB_iframe' => 'true',
+						'width'     => 270,
+						'height'    => 400,
+					),
+					self_admin_url( "theme-install.php" ) ) );
+			}
+
+			return $details_url;
+		}, 10, 2 );
 
 		/*
 		 * Update transient if necessary.
@@ -391,7 +360,6 @@ class Theme extends Base {
 		if ( empty( $current->response ) && empty( $current->up_to_date ) ) {
 			$this->pre_set_site_transient_update_themes( $current );
 		}
-
 
 		if ( isset( $current->up_to_date[ $theme_key ] ) ) {
 			$enclosure = $this->update_row_enclosure( $theme_key, 'theme' );
@@ -424,34 +392,6 @@ class Theme extends Base {
 				esc_html_e( 'No previous tags to rollback to.', 'github-updater' );
 			}
 			echo $enclosure['close'];
-		}
-
-		if ( isset( $current->response[ $theme_key ] ) ) {
-			$r = $current->response[ $theme_key ];
-			printf( esc_html__( 'GitHub Updater shows a new version of %s available.', 'github-updater' ),
-				$theme_name
-			);
-			printf( ' <a href="%s" class="thickbox" title="%s"> ',
-				$details_url,
-				$theme_name
-			);
-			if ( empty( $r['package'] ) ) {
-				printf( esc_html__( 'View version %s details.', 'github-updater' ),
-					$r['new_version']
-				);
-				echo '</a><em>';
-				esc_html_e( 'Automatic update is unavailable for this theme.', 'github-updater' );
-				echo '</em>';
-			} else {
-				printf( esc_html__( 'View version %1$s details%2$s or %3$supdate now%4$s.', 'github-updater' ),
-					$r['new_version'],
-					'</a>',
-					'<a href="' . wp_nonce_url( self_admin_url( 'update.php?action=upgrade-theme&theme=' ) . $theme_key, 'upgrade-theme_' . $theme_key ) . '">',
-					'</a>'
-				);
-			}
-
-			do_action( "in_theme_update_message-$theme_key", $theme, $r );
 		}
 	}
 
@@ -504,48 +444,6 @@ class Theme extends Base {
 		}
 		print( '</ul>' );
 		echo $enclosure['close'];
-	}
-
-	/**
-	 * Remove default after_theme_row_$stylesheet.
-	 *
-	 * @author @grappler
-	 *
-	 * @param $theme_key
-	 * @param $theme
-	 */
-	public function remove_after_theme_row( $theme_key, $theme ) {
-
-		foreach ( parent::$git_servers as $server ) {
-			$repo_header = $server . ' Theme URI';
-			$repo_uri    = $theme->get( $repo_header );
-
-			/**
-			 * Filter to add themes not containing appropriate header line.
-			 *
-			 * @since   5.4.0
-			 * @access  public
-			 *
-			 * @param   array $additions    Listing of themes to add.
-			 *                              Default null.
-			 * @param   array $theme        Current theme.
-			 * @param         string        'theme'    Type being passed.
-			 */
-			$additions = apply_filters( 'github_updater_additions', null, $theme, 'theme' );
-			foreach ( (array) $additions as $addition ) {
-				if ( $theme_key === $addition['slug'] ) {
-					if ( ! empty( $addition[ $server . ' Theme URI' ] ) ) {
-						$repo_uri = $addition[ $server . ' Theme URI' ];
-						break;
-					}
-				}
-			}
-			if ( empty( $repo_uri ) ) {
-				continue;
-			}
-
-			remove_action( "after_theme_row_$theme_key", 'wp_theme_update_row', 10 );
-		}
 	}
 
 	/**